requirements.txt
pyperclip==1.9.0
pytest==8.4.2
python-dotenv==1.1.1
cryptography>=41.0.0
